requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
watchdog>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pysrt>=1.1.2
chardet>=5.2.0
schedule>=1.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
aiohttp>=3.8.0
aiofiles>=23.0.0
